20150915 2122CST hours troubleshooting wAny and hAny issue.

+++ Bulltipeer/Bulltipeer/ConnectionsViewController.m
@@ -58,7 +58,7 @@
         }
         else if (state == MCSessionStateNotConnected){
             if ([_arrConnectedDevices count] > 0) {
-                int indexOfPeer = [_arrConnectedDevices indexOfObject:peerDisplayName];
+                long indexOfPeer = [_arrConnectedDevices indexOfObject:peerDisplayName];
                 [_arrConnectedDevices removeObjectAtIndex:indexOfPeer];
             }
